Fix for regression against BEVENT in handling "99" plays.
As of late 2009, BEVENT now no longer includes a putout or assist
in its output when fielding credits are specified as "99"
(the code for "unknown out").  cwevent had previously also
included these in the putout and assist fields for compatibility.
This change brings cwevent into line with BEVENT (and with behavior
that is more expected from end users!)