Enhancement: handle illegal substitution of P in DH game.
Under the current DH rule, pitchers may only bat for the DH and
enter the batting order in the "DH slot."  On at least one instance,
on 1976/9/5, this rule was violated, when Catfish Hunter entered as
a PH for another player.

Implements new subsitution logic to catch this possibility, and clear
out the slot used for the pitcher in DH games when this occurs.