Bugfix: Correct logic for GF pitcher in cwgame when team loses DH.
The previous logic in cwgame treated a pitcher entering the batting
order when a team lost the DH as a pitching substitution.
Correct this by instead computing the number of pitchers a team
uses, and reporting the last if there has been more than one.
This does still mean the GF pitcher can be the starter, if the
starter leaves the mound to play another position then returns.