Fix bug in SportsML generation when NP is first event.
Fix a bug, currently affecting only SportsML generation, in the edge case
where NP is the first event (e.g. because a substitution is recorded
prior to the first play).

This was occurring because of an incorrect initialization of state in
the game iterator, which failed to flag this as a situation in which
parsing of the event was successful.