Refactor game lint checking into separate function.
Remove some of the sanity checking previously done in constructing
a boxscore for a game, and moving it into a new function cw_game_lint().
This now allows a game to be checked for internal consistency without
terminating the program if an inconsistency is encountered.

Only some inconsistencies pertatining to event files are covered by
this revision.  Some errors in boxscore construction will still trigger
termination, pending consideration of ways to handle these gracefully.