Remove batted ball types from bunt strikeouts.
Plays which were coded as strikeouts on bunts previously generated a
batted ball type.  This matched the former behavior of BEVENT.
With the April 2013 version, strikeouts no longer generated batted
ball types; this revision brings cwevent into line.  This would seem
to be a sensible change in the interpretation of the batted ball
type output field.