Fix bug in incorrectly flagging errors as throwing.
When a /TH flag appeared in an advancement modifier, it was assumed that
it was modifying an error that previously appeared in that same modifier.
However, this assumption was not being checked, and it is legal for a /TH
modifier to appear at the end of a fielding credit that terminates in an
out and does not contain an error.  When this occurred, a previous unrelated
error in the play string would be flagged (incorrectly) as a throwing error.

This bug is now corrected by explicitly confirming the play in the advancement
modifier is a play string on which the runner was safe or advanced on an
error.