Improve boxscore generation in unusual re-entry cases.
When creating a boxscore from a full event file, the generator now
only searches players currently in the lineup.  This takes care of
the unusual case of a player leaving the game, and then re-entering
in a different lineup slot.  While this is illegal under current
professional rules, this can occur in courtesy runner situations, in
cases where the rules permit such re-entry, or where such re-entry
occurs even though it is illegal.