Correctly apply rule awarding games played at position.
Under modern rules, players only receive credit for a defensive game
played at a position if they appear there for at least one event.
This corrects the logic awarding the game played in boxscores.

Players will still be listed in the boxscore with the position, but
the 'g' field in the statistical record for the position will be zero
until the player has appeared there for at least one event.