Revise pitcher responsibility assignment on S/BR.
This changes the handling of pitcher responsibility assignment on S/BR plays.
Per the interpretation of the rule as promulgated by Elias, the baserunner
who is out on a S/BR is treated as if he were out taking an extra base, and
not due to "batter action."

For example, consider the following sequence:
* Abel hits a single off Pitcher Smith.
* Pitcher Thomas relieves Smith.
* Baker hits a S/BR, with Abel out.

In this circumstance, Baker is the responsibility of Thomas, and Smith
is now off the hook for the baserunner he left on in the person of Abel.

Previously, Chadwick has handled this as if it were a force out, which is
consistent with the scoring in the first instances of this found from the
1960s.