Implement international tiebreaker extension.
This commit implements an extension of the international tiebreaker
rule.  Depending on the variation, this rule places one or more baserunners
on base at the start of an inning.  In 2017 this is being used in
the complex-based affiliated leagues, although it has been previously
used in independent leagues as well as international competitions.

To place a runner on a base, the syntax is similar to the existing
badj and ladj records:

cw:itb,runner-id,base

This should be followed immediately by a NP event, as placing the runner
on base is conceptually its own event.

The runner is "charged" to the pitcher/catcher currently in the game for
the opposition.  By rule any run scored by this runner (or by a subsequent
runner who reaches due to this runner being put out) is unearned, but
as with all unearned runs, it is up to the event file to mark those
correctly with a (UR) notation.

Computation of runners left on base is updated to reflect these additional
runners, as with these runners the usual formula of LOB = plate
appearances - runs - outs no longer applies.

The record type has been prefixed by the cw: namespace to avoid
clashes with possible future support for this in the official
DiamondWare tools.  The use of this extension will render the
files incompatible with the current version of those tools.