Correct bug in computing day-of-week for pre-1900 dates.
The previous implementation for computing the day-of-week was
only valid post-1900.  This means it did not work for 19th century
games, and indeed could result in a core dump as it returned
negative values for the day of week.

This replaces the implementation with the one posted by Tomohiko
Sakamoto to comp.lang.c in 1992.