Refactor and abstract manipulation of runner data.
Over the years we have tracked more data about current baserunners,
and accessing this data has required knowledge of the implementation
details of how the data are stored.

This refactor reorganises the tracked data into a concept of a
'runner' structure, and all access to non-trivial queries or
manipulation of the structure is done through a suite of
functions that hide the implementation details.