Implement new advancement code for automatic runners.
A new advancement code 7 is now emitted in the case where a
run is scored by an automatic runner placed on base via an
radj record.

For the purposes of determining when the 7 code is used,
in the case where an automatic runner is put out by batter action,
the subsequent runner becomes considered as an automatic runner,
following the same rule used to assign responsibility for runs to
pitchers.