Add support for Ohtani DH rule
This adjusts the statistical tabulation for starting pitchers who DH
for themselves.  These players are considered to have two distinct
identities within the game for lineup and substitution purposes.
We adopt the convention that batting and non-pitcher fielding statistics
are tabulated against the boxscore entry in the batting order, and
pitching and pitcher fielding statistics against the non-batting-order
identity.