Remove coding of automatic runner scoring as advance type 7.
Provisionally in 2020, DiamondWare was going to code 'automatic' runners
who score as advance code 7.

However, this turns out to be ambiguous, as the 'automatic' runner
can turn out to be an earned run depending on the
scorer's discretion in reconstructing an inning when
assigning earned runs.

Meanwhile, DiamondWare has not implemented this code in published versions of BEVENT.

This commit reverts to using codes 4, 5, or 6 as appropriate irrespective
of whether the runner is the 'automatic' runner.

As a replacement for that functionality, this also implements new extended fields
which report automatic runner status, so the same information remains
available - but separates the concerns of earned/unearned run from whether the runner was the result
of the automatic runner placement.